% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_anndata_json.R
\name{json_to_seurat}
\alias{json_to_seurat}
\title{json_to_seurat}
\usage{
json_to_seurat(json_file, idents = NULL)
}
\arguments{
\item{json_file}{}
}
\value{
A Seurat object
}
\description{
Convert a JSON file into a Seurat object
}
