% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratJSON-package.R
\docType{package}
\name{seuratJSON-package}
\alias{seuratJSON}
\alias{seuratJSON-package}
\title{seuratJSON: Converts a JSON file into a Seurat object}
\description{
Reads an JSON file (currently, just serialized AnnData objects) and converts it into a Seurat object
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/milescsmith/seuratJSON}
  \item Report bugs at \url{https://github.com/milescsmith/seuratJSON/issues}
}

}
\keyword{internal}
